In SPDIF TX, fill port buffer before with zeroes before transmitting actual samples
This ensures that the sender is blocked trying to send data to the SPDIF thread and the
send happens as soon as the spdif TX is ready to receive the data (ideally, when the last port p <: word of the previous frame returns),
giving spdif TX enough time to produce the next frame that needs to be output on the pin.

Doing this fixes intermittent failures seen in the spdif out tests in sw_usb_audio where the
first preamble bit of a new frame misses timing.